requirements.txt
Django>=4.0
mysqlclient>=2.0
Pillow>=10.0